backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
schedule==1.2.0
python-dotenv==1.0.0